add a few checks when installing

+++ install.sh
@@ -4,7 +4,10 @@ DIR=$(dirname "$(readlink -m "$0")")
 
 sudo ubuntu-drivers autoinstall
 
+[[ -e ${HOME}/.bash_profile ]] && rm ~/.bash_profile
 cp "${DIR}"/.bash_profile_dist ~/.bash_profile
+
+[[ -e ${HOME}/.bashrc ]] && rm ~/.bashrc
 ln -s "${DIR}"/.bashrc ~/.bashrc
 source ~/.bashrc
 
@@ -12,10 +15,5 @@ git config --global user.name "Nelson da Costa"
 git config --global user.email "ndc@octo.com"
 git config --global include.path "${DIR}"/git/.gitconfig
 
-git clone https://github.com/michaeldfallen/git-radar ~/git-radar
-ln -s ~/git-radar/git-radar ~/bin/git-radar
-
-sudo apt install composer
-
-composer global require bamarni/symfony-console-autocomplete
-composer global require friendsofphp/php-cs-fixer
+[[ ! -d ${HOME}/git-radar ]] && git clone https://github.com/michaeldfallen/git-radar ~/git-radar
+[[ ! -e ${HOME}/bin/git-radar ]] && ln -s ~/git-radar/git-radar ~/bin/git-radar
